Chapter 02 Exercise 01
Read the list of files in the current directory and convert the names
to their full path specification. Don’t use the shell or an external
program to get the current directory. The File::Spec and Cwd modules,
both of which come with Perl, should help. Print each path with four
spaces before it and a newline after it.

This exercise was REALLY similar to chapter 11 exercise 1 of Learning
Perl, so it took me far less than the suggested 25 minutes. I wonder how
many of these will be repeats? Probably not that many.

/rs.