Chapter 02 Exercise 02
Install the local::lib module and use it when you install
Module::CoreList (or another module if you like). Write a program that
reports the name and first release date for all the modules in
Perl v5.14.2. Read the documentation for local::lib to see if it has
special installation instructions.

Okay, this time, the book spent more time making the output pretty than
I did. This is unusual. So I took their tricks and figured them out and
used them. :)

/rs.