Chapter 02 Exercise 03
Parse the International Standard Book Number from the back of this book
(9781449393090). Install the Business::ISBN module from CPAN and use it
to extract the group code and the publisher code from the number.

Interestingly, the answer key for this question uses
$isbn->country_code. This doesn't exist within the module, according to
both the documentation and the implementation. Neither does the question
itself ask for it.

/rs.