Chapter 03 Exercise 01
Write a program that takes a list of filenames on the command line and
uses grep to select the ones whose size is less than 1,000 bytes. Use
map to transform the strings in this list, putting four space characters
in front of each and a newline character after. Print the resulting list.

At the end of Learning Perl, I was having problems with the map
concept. This exercise pretty much solidified it for me, and grep as
well. I can see why they're so powerful!

/rs.