Chapter 03 Exercise 02
Write a program that asks the user to enter a regular expression. Read
this as data from standard input; don’t get it from the command line
arguments. Report a list of files in some hardcoded directory whose
names match the pattern. Repeat this until the user enters an empty
string instead of a pattern.

I had originally wrapped the entire assignment to @matching_files in the
eval, which lead me to have to define that array outside of the eval.
The answer key demonstrated that the eval could exist within the grep,
which I recall now from the text.

I also prefer testing for a newline with a regex, rather than the way
the answer key does it.

Finally, I have more robust error checking in my answer, and I actually
do something with the returned error. I also customize the output if
there's no matches.

/rs.