Chapter 04 Exercise 02
Using the final version of check_required_items, write a subroutine
check_items_for_all that takes as its only parameter a reference to a
hash whose keys are the people aboard the Minnow and whose corresponding
values are array references of the things they intend to bring on board.

It is taking a bit to keep the references straight in my head, but it's
starting to make sense again.

/rs.