Chapter 04 Exercise 03
Modify the crew roster program to add a location field for each castaway.
At the start, set each person’s location to “The Island.” After you’ve
added that field to each person’s hash, change the Howells’ locations
to “The Island Country Club.” Make a report of everyone’s location.

Still need more practice to remember which syntax is used in which
situation when dereferencing and whatnot. I'll get it eventually.

/rs.