Chapter 05 Exercise 02
Modify the code from this chapter that reads coconet.dat, so that each
source machine’s portion of the output shows the total bytes from that
machine. List the source machines in order from most to least data
transferred. Within each group, list the destination machines in order
from most to least data transferred to that target from the source
machine.

I've included the datafile, as I've added some extra data for testing.

I used an extra hash to keep track of the total traffic for each source,
rather than a hash within the total_bytes hash. I also skipped the
temporary array to hold the sources list.

I need to remember the levels of dereferencing, and where to put the
arrows when they're needed. This is going to take me a while to get
cemented I think.

Also, when I'm confused as to why my hash is trying to use all of the
data, instead of just the values, in its sort function, remind me to
actually CALL for keys or values when I'm looping over them in a for.

/rs.