Chapter 06 Exercise 01
Use the Storable module to modify the program from Exercise 2 in
Chapter 5 needs to keep the running totals in a data file so the
Professor can run it on each day’s logs to get the new totals

I ended up modifying the program from ex05-2.pl, because it was WAY
easier to do this with the data all in one place. I could likely have
done it with two, but this was just EASIER.

There's minimal error checking in place, and there's a lot of
improvements that could be done, also as per the discussion in the text,
but this answers the exercise as described.

I also included several data files that can be used as input, just for
the sake of convenience.

/rs.