Chapter 06 Exercise 02
Modify the program from Exercise 1 to use JSON instead of Storable.

An interesting problem occurred when I was doing this, and once I read
the solution, I think I understand what it was. In my implementation,
I'm reading the JSON data in with from_json / decode_json, and turning
the returned hash reference directly into a hash. That way, I didn't
have to change the rest of my code.

HOWEVER, when I 'pretty' print the JSON to the output file, I was
getting an error, choking right after the first character. The problem
was, I believe, that when it was a multi-line, pretty-printed file, it
was reading a line at a time, which was giving %total_bytes incorrect
data. But, when it was a single string, with no new-lines, it was
getting back data all the data at once, and was able to put it into the
hash immediately.

Or something like that.

/rs.