Chapter 08 Exercise 01
Write a program that prints the today’s date and the day of the week,
but allows the user to choose to send the output to a file, a scalar,
or both at the same time. No matter which one the user selects, send
the output with a single print statement. If the user chooses to send
the output to a scalar, at the end of the program print the scalar’s
value to standard output.

I used IO::File and IO::Scalar to open a file and a scalar, a given /
when to pick where to output, and several slices of localtime to get the
date just right, but it's basically the same as the book. I didn't use
the same filehandle to print, initially. Originally I had a different
filehandle for each, but that was just silly.

/rs.