Chapter 08 Exercise 02
The Professor has to read a logfile. He wants to write a series of
files, called gilligan.info, maryann.info, and so on. Each file should
contain all the lines that begin with that name. Now, the logfile is
large and the coconut-powered computer is not fast, so he wants to process
the input file in one pass and write all output files in parallel. How
does he do it?

... okay, my answer is WAY different than the books. More error
checking. Slurp mode. IO::Scalar. output_handles state variable in the
while.

Oh well, it all seems to work!

/rs.