Chapter 08 Exercise 03
Write a program that takes multiple directory names from the command
line and then prints out their contents. Use a function that takes a
directory handle reference that you made with opendir.

I understand what the book was doing by using 'grep' and 'map' and
several arrays, but I don't think it was terribly efficient to do it
that way. I like my way better, and it also works.

/rs.