Chapter 09 Exercise 01
Get the rightmost program running. Once you have the example working,
modify the right most program, take a hash reference of patterns, and
return the key of the rightmost match.

A quick reference to the docs for 'each' revealed that using it on the
hash reference would return the key instead of the array position. Using
this tidbit, I modified the rightmost function to record both the key
and the position. I could have returned them both in a list, I suppose,
and saved calling the original_rightmost function to get the position.

/rs.