Chapter 09 Exercise 02
Write a program to read in a list of patterns from a file. Precompile
the patterns and store them in an array. Prompt the user for lines of
input, printing the line number and text for each line that matches.
The $. variable is useful here.

Perldoc is again effective in explaining what '$.' is good for.

/rs.