Chapter 10 Exercise 01
Using the glob operator, a naïve sort of every name in your home
directory by their relative sizes might be written as:

    chdir;  # the default is our home directory
    my @sorted = sort { −s $a <=> −s $b } glob '*';

Rewrite this using the Schwartzian Transform technique.

We maintain the glob, since we need to get the file names from
somewhere. Next, we map the output of the glob onto an anonymous array
reference, along with the element that we'll be sorting by; that is, the
size of the current file.

Then, we sort by the sizes (element '1' in the anonymous array ref), and
again map the output back onto an array, this time element '0' (the
filename).

Confusing? A little. There's a lot of work happening in the Perl
interpreter here, that we aren't concerning ourselves with. We ahve to
trust that it's doing the right thing! Luckily, it is.

/rs.