Chapter 10 Exercise 03
Using a Schwartzian Transform, read a list of words and sort them in
“dictionary order.” Dictionary order ignores all capitalization and
internal punctuation. Be sure you don’t mangle the data! If the input
includes the Professor and The Skipper, the output should have them listed
in that order, with that capitalization.

I liked this exercise. It really helped to start cementing the
usefulness of map and sort, and their syntax. I also went through a
couple of iterations of the transform; I started by normalizing all of
the strings into their own array, and doing the sort based on array
index. I realized, though, that I could just normalize the strings right
in the initial map, and it simplified things immensely.

/rs.