Chapter 10 Exercise 04
Modify the recursive directory dumping routine so it shows the nested
directories through indentation. An empty directory should show up as:

sandbar, an empty directory

while a nonempty directory should appear with nested contents, indented
two spaces.

I had way more trouble with this one. The sample code was wrong in the
book, so it took me forever to debug it. Then I couldn't figure out how
to determine if a directory was empty or not... I had to look at the
solution. I got the indentation right, though!

/rs.