Chapter 11 Exercise 01
The Oogaboogoo natives on the island have unusual names for the days and
months. Here is some simple but not very well-written code from Gilligan.
Fix it up, add a conversion function for the month names, and make the
whole thing into a library. For extra credit, add suitable error checking
and consider what should be in the documentation.

I allowed the user to specify 7 for 'ark', since we allow them to do
that often in Julian dates.

/rs.