Chapter 13, Exercise 2
Create a program to ask the user to enter the names of one or more
barnyard animals. Create a barnyard with those animals, and have each
animal speak once.

An interesting lesson in the need for tests arose in this one. My
'mouse' class had an error in it, where I was calling the superclass
like this:

	class->SUPER::speak( $class )

Spot the error? It took me a few minutes to see that I'd forgotten the
'$' prefix on the class. A bit of debugging (after it crashed not able
to find 'class', which also threw me for a loop), I found the problem,
but if I had proper tests when I started, that would have gotten caught
quite quickly!

/rs.